backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
Pillow==10.0.1
ffmpeg-python==0.2.0
marshmallow==3.20.1
webargs==8.3.0
gunicorn==21.2.0
python-multipart==0.0.6
email-validator==2.1.0
